requirements.txt
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
fonttools==4.52.1
h5py==3.11.0
imageio==2.34.1
imgaug==0.4.0
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
psutil==5.9.8
pyclipper==1.3.0.post5
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scikit-image==0.23.2
scipy==1.13.1
shapely==2.0.4
six==1.16.0
tifffile==2024.5.22
tqdm==4.66.4
validators==0.28.2
